+++ fiware/CMakeLists.txt
@@ -18,6 +18,7 @@ cmake_minimum_required(VERSION 3.5.0)
 
 project(soss-fiware)
 
+find_package(fastrtps REQUIRED)
 find_package(soss-core REQUIRED)
 
 if(WIN32)
@@ -66,11 +67,7 @@ endif()
 ########################################################################
 # Library
 set(SRC
-    src/Conversion.cpp
-    src/Listener.cpp
     src/NGSIV2Connector.cpp
-    src/Publisher.cpp
-    src/Subscriber.cpp
     src/SystemHandle.cpp
 )
