--- fiware/src/SystemHandle.hpp
@@ -1,72 +1,0 @@
-﻿/*
- * Copyright 2019 Proyectos y Sistemas de Mantenimiento SL (eProsima).
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
-*/
-
-#ifndef SOSS__FIWARE__INTERNAL__SYSTEMHANDLE_HPP
-#define SOSS__FIWARE__INTERNAL__SYSTEMHANDLE_HPP
-
-#include "Publisher.hpp"
-#include "Subscriber.hpp"
-
-#include <soss/SystemHandle.hpp>
-#include <soss/core/export.hpp>
-
-#include <vector>
-#include <memory>
-
-namespace soss {
-namespace fiware {
-
-
-class NGSIV2Connector;
-
-class SystemHandle : public virtual TopicSystem
-{
-public:
-    SystemHandle() = default;
-    virtual ~SystemHandle() override;
-
-    bool configure(
-        const RequiredTypes& types,
-        const YAML::Node& configuration) override;
-
-    bool okay() const override;
-
-    bool spin_once() override;
-
-    bool subscribe(
-        const std::string& topic_name,
-        const std::string& message_type,
-        SubscriptionCallback callback,
-        const YAML::Node& configuration) override;
-
-    std::shared_ptr<TopicPublisher> advertise(
-        const std::string& topic_name,
-        const std::string& message_type,
-        const YAML::Node& configuration) override;
-
-private:
-    std::unique_ptr<NGSIV2Connector> fiware_connector_;
-    std::vector<std::shared_ptr<Publisher>> publishers_;
-    std::vector<std::shared_ptr<Subscriber>> subscribers_;
-};
-
-
-} // namespace fiware
-} // namespace soss
-
-
-#endif // SOSS__FIWARE__INTERNAL__SYSTEMHANDLE_HPP
